+++ behave/configuration.py
@@ -4,11 +4,12 @@ from __future__ import print_function
 import os
 import re
 import sys
-import argparse
 import logging
 import shlex
 import six
+import argparse
 from six.moves import configparser
+from six import string_types
 
 from behave.model import FileLocation, ScenarioOutline
 from behave.reporter.junit import JUnitReporter
@@ -16,10 +17,13 @@ from behave.reporter.summary import SummaryReporter
 from behave.tag_expression import TagExpression
 from behave.formatter.base import StreamOpener
 from behave.formatter.formatters import formatters as registered_formatters
+from behave.userdata import UserData, parse_user_define
+from behave._types import Unknown
 
 
-class Unknown(object): pass
-
+# -----------------------------------------------------------------------------
+# CONFIGURATION DATA TYPES:
+# -----------------------------------------------------------------------------
 class LogLevel(object):
     names = [
          "NOTSET", "CRITICAL", "FATAL", "ERROR",
@@ -55,6 +59,11 @@ class ConfigError(Exception):
     pass
 
 
+
+
+# -----------------------------------------------------------------------------
+# CONFIGURATION SCHEMA:
+# -----------------------------------------------------------------------------
 options = [
     (('-c', '--no-color'),
      dict(action='store_false', dest='color',
@@ -69,6 +78,11 @@ options = [
     (('-d', '--dry-run'),
      dict(action='store_true',
           help="Invokes formatters without executing the steps.")),
+
+    (('-D', '--define'),
+     dict(dest='userdata_defines', type=parse_user_define, action='append',
+          help="""Define user-specific data in config.userdata dictionary.
+                  Example: -D foo=bar to store it in config.userdata["foo"].""")),
 
     (('-e', '--exclude'),
      dict(metavar="PATTERN", dest='exclude_re',
@@ -106,6 +120,11 @@ options = [
           help="""Specify a formatter. If none is specified the default
                   formatter is used. Pass '--format help' to get a
                   list of available formatters.""")),
+
+    (('--steps-catalog',),
+     dict(action='store_true', dest='steps_catalog',
+          help="""Show a catalog of all available step definitions.
+                  SAME AS: --format=steps.catalog --dry-run --no-summary -q""")),
 
     ((),  # -- CONFIGFILE only
      dict(dest='scenario_outline_annotation_schema',
@@ -338,6 +357,7 @@ raw_value_options = frozenset([
     # -- MAYBE: "scenario_outline_annotation_schema",
 ])
 
+
 def read_configuration(path):
     cfg = configparser.SafeConfigParser()
     cfg.read(path)
@@ -364,6 +384,8 @@ def read_configuration(path):
         elif action in ('store_true', 'store_false'):
             result[dest] = cfg.getboolean('behave', dest)
         elif action == 'append':
+            if dest == 'userdata_defines':
+                continue    # -- SKIP-CONFIGFILE: Command-line only option.
             result[dest] = \
                 [s.strip() for s in cfg.get('behave', dest).splitlines()]
         else:
@@ -393,21 +415,29 @@ def read_configuration(path):
             result[paths_name] = \
                 [os.path.normpath(os.path.join(cfgdir, p)) for p in paths]
 
+    result['userdata'] = {}
+    if cfg.has_section('behave.userdata'):
+        result['userdata'].update(cfg.items('behave.userdata'))
     return result
 
 
-def load_configuration(defaults, verbose=False):
+def config_filenames():
     paths = ['./', os.path.expanduser('~')]
     if sys.platform in ('cygwin', 'win32') and 'APPDATA' in os.environ:
         paths.append(os.path.join(os.environ['APPDATA']))
 
-    for path in paths:
-        for filename in 'behave.ini .behaverc'.split():
+    for path in reversed(paths):
+        for filename in reversed(('behave.ini', '.behaverc', 'setup.cfg')):
             filename = os.path.join(path, filename)
             if os.path.isfile(filename):
-                if verbose:
-                    print('Loading config defaults from "%s"' % filename)
-                defaults.update(read_configuration(filename))
+                yield filename
+
+
+def load_configuration(defaults, verbose=False):
+    for filename in config_filenames():
+        if verbose:
+            print('Loading config defaults from "%s"' % filename)
+        defaults.update(read_configuration(filename))
 
     if verbose:
         print('Using defaults:')
@@ -415,29 +445,30 @@ def load_configuration(defaults, verbose=False):
             print('%15s %s' % (k, v))
 
 
-# construct the parser
-#usage = "%(prog)s [options] [ [FILE|DIR|URL][:LINE[:LINE]*] ]+"
-usage = "%(prog)s [options] [ [DIR|FILE|FILE:LINE] ]+"
-description = """\
-Run a number of feature tests with behave."""
-more = """
-EXAMPLES:
-behave features/
-behave features/one.feature features/two.feature
-behave features/one.feature:10
-behave @features.txt
-"""
-parser = argparse.ArgumentParser(usage=usage, description=description)
-for fixed, keywords in options:
-    if not fixed:
-        continue    # -- CONFIGFILE only.
-    if 'config_help' in keywords:
-        keywords = dict(keywords)
-        del keywords['config_help']
-    parser.add_argument(*fixed, **keywords)
-parser.add_argument('paths', nargs='*',
+def setup_parser():
+    # construct the parser
+    #usage = "%(prog)s [options] [ [FILE|DIR|URL][:LINE[:LINE]*] ]+"
+    usage = "%(prog)s [options] [ [DIR|FILE|FILE:LINE] ]+"
+    description = """\
+    Run a number of feature tests with behave."""
+    more = """
+    EXAMPLES:
+    behave features/
+    behave features/one.feature features/two.feature
+    behave features/one.feature:10
+    behave @features.txt
+    """
+    parser = argparse.ArgumentParser(usage=usage, description=description)
+    for fixed, keywords in options:
+        if not fixed:
+            continue    # -- CONFIGFILE only.
+        if 'config_help' in keywords:
+            keywords = dict(keywords)
+            del keywords['config_help']
+        parser.add_argument(*fixed, **keywords)
+    parser.add_argument('paths', nargs='*',
                 help='Feature directory, file or file location (FILE:LINE).')
-
+    return parser
 
 class Configuration(object):
     defaults = dict(
@@ -452,13 +483,16 @@ class Configuration(object):
         log_capture=True,
         logging_format='%(levelname)s:%(name)s:%(message)s',
         logging_level=logging.INFO,
+        steps_catalog=False,
         summary=True,
         junit=False,
         stage=None,
+        userdata={},
         # -- SPECIAL:
         default_format="pretty",   # -- Used when no formatters are configured.
         scenario_outline_annotation_schema=u"{name} -- @{row.id} {examples.name}"
     )
+    cmdline_only_options = set("userdata_defines")
 
     def __init__(self, command_args=None, load_config=True, verbose=None,
                  **kwargs):
@@ -500,12 +534,14 @@ class Configuration(object):
         self.scenario_outline_annotation_schema = None
         self.steps_dir = "steps"
         self.environment_file = "environment.py"
+        self.userdata_defines = None
         if load_config:
             load_configuration(self.defaults, verbose=verbose)
+        parser = setup_parser()
         parser.set_defaults(**self.defaults)
         args = parser.parse_args(command_args)
         for key, value in args.__dict__.items():
-            if key.startswith('_'):
+            if key.startswith('_') and key not in self.cmdline_only_options:
                 continue
             setattr(self, key, value)
 
@@ -518,6 +554,15 @@ class Configuration(object):
                     self.outputs.append(StreamOpener(outfile))
                 else:
                     self.outputs.append(StreamOpener(stream=sys.stdout))
+
+        if self.steps_catalog:
+            # -- SHOW STEP-CATALOG: As step summary.
+            self.default_format = "steps.catalog"
+            self.format = ["steps.catalog"]
+            self.dry_run = True
+            self.summary = False
+            self.show_skipped = False
+            self.quiet = True
 
         if self.wip:
             # Only run scenarios tagged with "wip".
@@ -565,6 +610,7 @@ class Configuration(object):
             self.stage = os.environ.get("BEHAVE_STAGE", None)
         self.setup_stage(self.stage)
         self.setup_model()
+        self.setup_userdata()
 
     def collect_unknown_formats(self):
         unknown_formats = []
@@ -605,8 +651,8 @@ class Configuration(object):
         But note that the logging setup can only occur once.
 
         SETUP MODES:
-          * :func:`logging.config.fileConfig()`, if ``configfile` is provided.
-          * :func:`logging.basicConfig(), otherwise.
+          * :func:`logging.config.fileConfig()`, if ``configfile`` is provided.
+          * :func:`logging.basicConfig()`, otherwise.
 
         .. code-block: python
             # -- FILE: features/environment.py
@@ -665,3 +711,19 @@ class Configuration(object):
             environment_file = prefix + environment_file
         self.steps_dir = steps_dir
         self.environment_file = environment_file
+
+    def setup_userdata(self):
+        if not isinstance(self.userdata, UserData):
+            self.userdata = UserData(self.userdata)
+        if self.userdata_defines:
+            # -- ENSURE: Cmd-line overrides configuration file parameters.
+            self.userdata.update(self.userdata_defines)
+
+    def update_userdata(self, data):
+        """Update userdata with data and reapply userdata defines (cmdline).
+        :param data:  Provides (partial) userdata (as dict)
+        """
+        self.userdata.update(data)
+        if self.userdata_defines:
+            # -- REAPPLY: Cmd-line defines (override configuration file data).
+            self.userdata.update(self.userdata_defines)