+++ behave/model.py
@@ -7,10 +7,11 @@ import itertools
 import logging
 import os.path
 import six
-from six import text_type as unicode
 import sys
 import time
 import traceback
+from six import string_types
+from six import text_type as unicode
 from behave import step_registry
 from behave.compat.os_path import relpath
 
@@ -74,7 +75,7 @@ class FileLocation(object):
         self.line = line
 
     # def __new__(cls, filename, line=None):
-    #     assert isinstance(filename, six.string_types)
+    #     assert isinstance(filename, string_types)
     #     obj = unicode.__new__(cls, filename)
     #     obj.line = line
     #     obj.__filename = filename
@@ -112,7 +113,7 @@ class FileLocation(object):
     def __eq__(self, other):
         if isinstance(other, FileLocation):
             return self.filename == other.filename and self.line == other.line
-        elif isinstance(other, six.string_types):
+        elif isinstance(other, string_types):
             return self.filename == other
         else:
             raise AttributeError("Cannot compare FileLocation with %s:%s" % \
@@ -139,7 +140,7 @@ class FileLocation(object):
                 else:
                     return self.line < other.line
 
-        elif isinstance(other, six.string_types):
+        elif isinstance(other, string_types):
             return self.filename < other
         else:
             raise AttributeError("Cannot compare FileLocation with %s:%s" % \