+++ behave/textutil.py
@@ -2,7 +2,8 @@
 """
 Provides some utility functions related to text processing.
 """
-import six
+
+from six import string_types, text_type
 
 
 def make_indentation(indent_size, part=u" "):
@@ -22,12 +23,12 @@ def indent(text, prefix):
     """
     lines = text
     newline = u""
-    if isinstance(text, six.string_types):
+    if isinstance(text, string_types):
         lines = text.splitlines(True)
     elif lines and not lines[0].endswith("\n"):
         # -- TEXT LINES: Without trailing new-line.
         newline = u"\n"
-    return newline.join([prefix + six.text_type(line) for line in lines])
+    return newline.join([prefix + text_type(line) for line in lines])
 
 
 def compute_words_maxsize(words):